feat(condo): DOMA-8258 united migrations
