+++ apps/condo/domains/billing/access/SendNewBillingReceiptFilesNotificationsService.js
@@ -1,5 +1,5 @@
 /**
- * Generated by `createservice organization.SendNewBillingReceiptFilesNotificationsService '--type=mutations'`
+ * Generated by `createservice billing.SendNewBillingReceiptFilesNotificationsService '--type=mutations'`
  */
 const { throwAuthenticationError } = require('@open-condo/keystone/apolloErrorFormatter')
