+++ apps/condo/domains/billing/tasks/index.js
@@ -1,0 +1,7 @@
+const { createTask } = require('@open-condo/keystone/tasks')
+
+const { sendNewBillingReceiptFilesNotifications } = require('./sendNewBillingReceiptFilesNotifications')
+
+module.exports = {
+    sendNewBillingReceiptFilesNotifications: createTask('sendNewBillingReceiptFilesNotifications', sendNewBillingReceiptFilesNotifications),
+}
