+++ apps/condo/domains/billing/utils/serverSchema/index.js
@@ -25,6 +25,7 @@ const { BillingCategory: BillingCategoryGQL } = require('@condo/domains/billing/
 const { REGISTER_BILLING_RECEIPTS_MUTATION } = require('@condo/domains/billing/gql')
 const { BillingReceiptFile: BillingReceiptFileGQL } = require('@condo/domains/billing/gql')
 const { VALIDATE_QRCODE_MUTATION } = require('@condo/domains/billing/gql')
+const { SEND_NEW_BILLING_RECEIPT_FILES_NOTIFICATIONS_MUTATION } = require('@condo/domains/billing/gql')
 const { SEND_RESIDENT_MESSAGE_MUTATION } = require('@condo/domains/resident/gql')
 /* AUTOGENERATE MARKER <IMPORT> */
 
@@ -108,6 +109,19 @@ async function validateQRCode (context, data) {
     })
 }
 
+async function sendNewBillingReceiptFilesNotifications (context, data) {
+    if (!context) throw new Error('no context')
+    if (!data) throw new Error('no data')
+    if (!data.sender) throw new Error('no data.sender')
+
+    return await execGqlWithoutAccess(context, {
+        query: SEND_NEW_BILLING_RECEIPT_FILES_NOTIFICATIONS_MUTATION,
+        variables: { data: { dv: 1, ...data } },
+        errorMessage: '[error] Unable to sendNewBillingReceiptFilesNotifications',
+        dataPath: 'result',
+    })
+}
+
 /* AUTOGENERATE MARKER <CONST> */
 
 module.exports = {
@@ -127,5 +141,6 @@ module.exports = {
     sendNewReceiptMessagesToResidentScopes,
     BillingReceiptFile,
     validateQRCode,
+    sendNewBillingReceiptFilesNotifications,
 /* AUTOGENERATE MARKER <EXPORTS> */
 }
